% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plane_geometry_utils.R
\name{EllipseFromCenterAndMatrix}
\alias{EllipseFromCenterAndMatrix}
\title{Ellipse from center and matrix}
\usage{
EllipseFromCenterAndMatrix(center, S)
}
\description{
Ellipse from center and matrix
}
