% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eoifunc.R
\name{eoifunc}
\alias{eoifunc}
\title{Ellipse of Insignificance calculator for 2x2 trials}
\usage{
eoifunc(a, b, c, d, v = 0.05, vis = 1, ssvec = c(NA, NA, NA, NA))
}
\arguments{
\item{a}{Number of positive cases in the experimental arm}

\item{b}{Number of negative cases in the experimental arm}

\item{c}{Number of positive cases in the control arm}

\item{d}{Number of negative cases in the control arm}

\item{v}{Optional alpha level to be tested (default p < 0.05)}

\item{vis}{Graphic summary on / off}

\item{ssvev}{An optional vector of experimental sensitivity, specificity and control sensitivity and specificity}
}
\value{
Ellipse of insignificance statistics and refined fragility measures for any 2x2 binary outcome trial
}
\description{
Ellipse of insignificance (EOI) analysis is a powerful refined fragility index to ascertaining robustness of results in dichotomous outcome trials, quantifying the degree of re-coding in both trial arms is required to lose (or gain) significance at a given level. It is powerful and flexible, and if the sensitivity and specificity data for a trial are also known, it can deduce whether reported results are robust. For full methods and citations, please see Grimes 2022 (Elife): "The ellipse of insignificance, a refined fragility index for ascertaining robustness of results in dichotomous outcome trials" - 10.7554/eLife.79573
}
\examples{
g <- eoifunc(700,300,500,500)
g <- eoifunc(113,887,24,976, ssvec = c(0.75,0.9,0.68,0.99))
}
