% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eoimeta2.R
\name{eoimeta}
\alias{eoimeta}
\title{Ellipse of Insignificance calculator for Meta Analyses of 2x2 trials}
\arguments{
\item{av}{List of positive cases in the experimental arm of individual trials}

\item{bv}{List of non-events in the experimental arm of individual trials}

\item{cv}{List of positive cases in the control arm of individual trials}

\item{dv}{List of non-events in the control arm of individual trials}

\item{v}{Optional alpha level to be tested (default p < 0.05)}

\item{atal}{Run Atal et al study specific fragility algorithm (default = 0, 1 for on)}

\item{wiv}{Used generic weighed inverse variance correction for included studies (default = 1)}
}
\value{
Ellipse of insignificance statistics and refined fragility measures for any 2x2 binary outcome trial
}
\description{
Extension of Ellipse of insignificance (EOI) and Region of Attainable Redaction (ROAR) analysis meta for 2x2 trials to meta-analysis.
}
