% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plane_geometry_utils.R
\name{radicalCenter}
\alias{radicalCenter}
\title{Radical center}
\usage{
radicalCenter(circ1, circ2, circ3)
}
\arguments{
\item{circ1, circ2, circ3}{\code{Circle} objects}
}
\description{
Radical center
}
