% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roarfunc.R
\name{roarfunc}
\alias{roarfunc}
\title{Region of Attainable redaction Analysis for 2x2 trials}
\usage{
roarfunc(a, b, c, d, v = 0.05, vis = 1)
}
\arguments{
\item{a}{Number of positive cases in the experimental arm}

\item{b}{Number of negative cases in the experimental arm}

\item{c}{Number of positive cases in the control arm}

\item{d}{Number of negative cases in the control arm}

\item{v}{Optional alpha level to be tested (default p < 0.05)}

\item{vis}{Graphic summary on / off}
}
\value{
ROAR statistics and hypothetical redaction levels for any 2x2 binary outcome trial
}
\description{
Region of Attainable redaction (ROAR) Analysis is an extention of Ellipse of Insignificance analysis for binary 2x2 trials that calculates the degree of redaction required to change a result from insignificant to significant or vice versa. It allows estimation of the degree of 'hidden' results required in control and experimental arms to change from null to non-null. For full methods and citations, please see Grimes 2024 (Elife): "Region of Attainable Redaction, an extension of Ellipse of Insignificance analysis for gauging impacts of data redaction in dichotomous outcome trials " - 10.7554/eLife.93050
}
\examples{
h <- roarfunc(700,300,500,500)
h <- roarfunc(250,550,400,500, v = 0.01)
}
