% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plane_geometry_utils.R
\docType{data}
\name{unitCircle}
\alias{unitCircle}
\title{Unit circle}
\format{
An object of class \code{Circle} (inherits from \code{R6}) of length 25.
}
\usage{
unitCircle
}
\description{
Unit circle
}
\keyword{datasets}
